Move forecast-related methods into Forecastable module
