+++ app/facades/road_trip_facade.rb
@@ -1,9 +1,14 @@
+require './lib/modules/forecastable'
+
 class RoadTripFacade
+  include Forecastable
+
   attr_reader :id, :origin, :destination, :travel_time, :arrival_forecast
 
   def initialize(origin, destination)
     @origin = origin
     @destination = destination
+    @geo_location = destination
   end
 
   def arrival_forecast
@@ -15,6 +20,7 @@ class RoadTripFacade
   end
 
   private
+
     def travel_time_seconds
       @travel_time_seconds ||= MapService.get_travel_time(@origin, @destination)
     end
@@ -43,30 +49,10 @@ class RoadTripFacade
     end
 
     def get_nearest_forecast
-      if hours <= 48
-        hourly.min_by{ |hour| (hour.unix_time - arrival_time).abs }
+      if hours <= 60
+        hourly_data.min_by{ |hour| (hour.unix_time - arrival_time).abs }
       else
-        daily.min_by{ |day| (day.unix_time - arrival_time).abs }
+        daily_data.min_by{ |day| (day.unix_time - arrival_time).abs }
       end
-    end
-
-    def geocode_data
-      @geo_data ||= MapService.get_coordinates(@destination)
-    end
-
-    def coordinates
-      geocode_data[:geometry][:location]
-    end
-
-    def forecast_data
-      @weather_data ||= OpenWeatherService.get_weather_data(coordinates[:lat], coordinates[:lng])
-    end
-
-    def hourly
-      @hourly ||= forecast_data[:hourly].map { |hour| HourlyForecast.new(hour, forecast_data[:timezone]) }
-    end
-
-    def daily
-      @daily ||= forecast_data[:daily].map { |day| DailyForecast.new(day, forecast_data[:timezone]) }
     end
 end
